Extract suffix generation and validation to a module
